Creando CRUD a las tablas para mantenimiento

+++ src/page/PageInventario/Inventario.jsx
@@ -22,16 +22,19 @@ function Inventario() {
   };
 
   const handleDelete = async (id) => {
-    try {
-      const response = await fetch(`http://localhost:3000/productos/${id}`, { method: 'DELETE' });
-      if (response.ok) {
-        alert('Producto eliminado');
-        fetchProductos();
+    if (window.confirm("¿Estás seguro de eliminar este producto?")) {
+      try {
+        const response = await fetch(`http://localhost:3000/productos/${id}`, { method: 'DELETE' });
+        if (response.ok) {
+          alert('Producto eliminado');
+          fetchProductos();
+        }
+      } catch (error) {
+        console.error('Error al eliminar el producto:', error);
       }
-    } catch (error) {
-      console.error('Error al eliminar el producto:', error);
     }
   };
+  
 
   const handleEdit = (producto) => {
     setProductoEdit(producto);
